Implémentation du score du capteur par salle , avec explication du calcul dans le front

+++ Back/src/main/java/esiag/back/controllers/capteur/CapteurController.java
@@ -31,7 +31,7 @@ public class CapteurController {
     }
 
     @GetMapping("salle")
-    public ResponseEntity<List<Capteur>> findBySalle(@RequestParam Integer idSalle){
+    public ResponseEntity<List<Capteur>> findBySalle(@RequestParam Long idSalle ){
         return new ResponseEntity<>(capteurService.findBySalle(idSalle), HttpStatus.OK);
     }
